\name{ffi-package}
\alias{ffi-package}
\alias{ffi}
\docType{package}
\title{
  Foreign Function Interface
}
\description{
  The \code{ffi} package is a replacement for the standard \code{.Call},
  \code{.C} and \code{.External} foreign function routines. In
  particular, it allows for the more extension specification and
  coercion of R types to foreign types.
}
\details{
}
\author{
  Byron Ellis \email{byron.ellis@gmail.com}
}
\keyword{package}